+++ tests/Main.hs
@@ -34,11 +34,12 @@
 
 module Main (main) where
 
+import Data.List.NonEmpty (NonEmpty (..))
 import Test.Hspec
 import Test.Hspec.Megaparsec
 import Text.Megaparsec
-import Text.Megaparsec.Error (newErrorMessages)
-import Text.Megaparsec.Pos (initialPos)
+import Text.Megaparsec.String
+import qualified Data.Set as E
 
 #if !MIN_VERSION_base(4,8,0)
 import Control.Applicative ((<*))
@@ -50,29 +51,30 @@ main :: IO ()
 main = hspec $ do
   describe "shouldParse" $
     it "works" $
-      parse letterChar "" "x" `shouldParse` 'x'
+      parse (letterChar :: Parser Char) "" "x" `shouldParse` 'x'
   describe "parseSatisfies" $
     it "works" $
-      parse (many punctuationChar) "" "?!!" `parseSatisfies` ((== 3) . length)
+      parse (many punctuationChar :: Parser String) "" "?!!"
+        `parseSatisfies` ((== 3) . length)
   describe "shouldFailOn" $
     it "works" $
-      parse (char 'x') "" `shouldFailOn` "a"
+      parse (char 'x' :: Parser Char) "" `shouldFailOn` "a"
   describe "shouldSucceedOn" $
     it "works" $
-      parse (char 'x') "" `shouldSucceedOn` "x"
+      parse (char 'x' :: Parser Char) "" `shouldSucceedOn` "x"
   describe "shouldFailWith" $
     it "works" $
-      parse (char 'x') "" "b" `shouldFailWith`
-        newErrorMessages [Unexpected "'b'", Expected "'x'"] (initialPos "")
+      parse (char 'x' :: Parser Char) "" "b" `shouldFailWith`
+        ParseError
+          { errorPos        = initialPos "" :| []
+          , errorUnexpected = E.singleton (Tokens $ 'b' :| [])
+          , errorExpected   = E.singleton (Tokens $ 'x' :| [])
+          , errorCustom     = E.empty }
   describe "failsLeaving" $
     it "works" $
-      runParser' (many (char 'x') <* eof) (initialState "xxa")
-#if MIN_VERSION_megaparsec(4,4,0)
+      runParser' (many (char 'x') <* eof :: Parser String) (initialState "xxa")
         `failsLeaving` "a"
-#else
-        `failsLeaving` "xxa"
-#endif
   describe "succeedsLeaving" $
     it "works" $
-      runParser' (many (char 'x')) (initialState "xxa")
+      runParser' (many (char 'x') :: Parser String) (initialState "xxa")
         `succeedsLeaving` "a"