+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Hspec Megaparsec 0.2.0
+
+* This version of `hspec-megaparsec` should be used with Megaparsec 5.
+
 ## Hspec Megaparsec 0.1.1
 
 * Make it pass tests with Megaparsec 4.4.0 and later.